Rewrote how ISSNs and Links were looked up in the browse results.
Rewrote how ISSNs and Links were looked up in the browse results. By
batch calling them before going to templates it cuts the number of query
requests by 100 and is twice as fast on a local machine. Should be even
more savings on a remote DB setup. Main code rev. 1771